Refactor to give things more structure
